% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isEqual.R
\name{isEqual}
\alias{isEqual}
\title{Verify equality}
\usage{
isEqual(x, y)
}
\arguments{
\item{x, }{object 1 to compare (vector, \code{matrix}, \code{data.frame})}

\item{y, }{object 2 to compare (vector, \code{matrix}, \code{data.frame})}
}
\value{
If \code{x} and \code{y} are vectors, a \code{logical} vector of the same length is returned; if \code{x} and \code{y} are of class \code{matrix} or \code{data.frame}, a \code{logical} matrix of the same dimension is returned.
}
\description{
Verify equality, including \code{NA} presence.
}
\details{
The length or dimension of \code{x} and \code{y} must be the same. The function rely on the \code{==} operator; if \code{x} or \code{y} are \code{numeric} with several significative numbers, one should use the function \code{isAlmostEqual} instead.
}
\examples{
x <- c(1:3,NA)
y <- c("1","C",NA, NA)
isEqual(x, y)
}
