% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNumericAnInteger.R
\name{isNumericAnInteger}
\alias{isNumericAnInteger}
\title{Identify if numbers are integers}
\usage{
isNumericAnInteger(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{vector of class \code{numeric}.}

\item{tol}{tolerance used when considering numeric values equal.}
}
\value{
Return a \code{logical} vector indicating for each element if it correspond to an integer.
}
\description{
Identify if elements of a \code{numeric} vector are integers.
}
\details{
The function also return \code{FALSE} for \code{NA}. Code in part taken from the example of the function \code{integer}.
}
\examples{
x <- c(1, 1.1, 0.1, 1.0, 1.00000001, NA)
isNumericAnInteger(x)

}
