% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isStringADate.R
\name{isStringADate}
\alias{isStringADate}
\title{Identify valid dates}
\usage{
isStringADate(x, format = ("\%Y-\%m-\%d"))
}
\arguments{
\item{x}{\code{character} vector.}

\item{format}{\code{character} representing a date (see \code{as.Date}). The only implemented format is \code{"\%Y-\%m-\%d"}.}
}
\value{
Return a \code{logical} vector identifying, for each element, if the string correspond to a valid date in the given format.
}
\description{
Identify if elements of a character vector represent valid dates in the designated format.
}
\examples{
x <- c("2019-06-20", "2019-06-32", "43636", NA, "2019/06/20")
isStringADate(x)

}
