% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isStringADecimal.R
\name{isStringADecimal}
\alias{isStringADecimal}
\title{Identify if a string is a decimal number}
\usage{
isStringADecimal(x, decimalSeparator = ".")
}
\arguments{
\item{x}{\code{character} vector.}

\item{decimalSeparator}{\code{character} used as a decimal separator ("," or "."); \code{decimalSeparator = "."} by default.}
}
\value{
Return a \code{logical} vector indicating, for each element of \code{x}, if the string correspond to a decimal number.
}
\description{
Check if elements of a \code{character} vector are decimal numbers (i.e. separated by a decimal symbol).
}
\details{
The function does not identify number in their scientific notation. Use the function \code{isStringSciNumber} for this purpose.
}
\examples{
x <- c("12.34", "12.00", "-1.1", "12,34", "-1.1E-2", "1234", "-1", "1E-2", "123A", "12.00.00", ".1")
isStringADecimal(x,".")
isStringADecimal(x,",")
}
