% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isStringAnInteger.R
\name{isStringAnInteger}
\alias{isStringAnInteger}
\title{Identify if a string is an integer}
\usage{
isStringAnInteger(x)
}
\arguments{
\item{x}{\code{character} vector}
}
\value{
Return a \code{logical} vector indicating, for each element of \code{x}, if the string correspond to an integer.
}
\description{
Check if elements of a \code{character} vector are integers
}
\details{
Use the function \code{isStringSciNumber} to identify numbers in their scientific notation.
}
\examples{
x <- c("1234", "-1","+1", "12.34","12,34", "123A","-1.1","1E+1")
isStringAnInteger(x)
}
