% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isStringANumber.R
\name{isStringANumber}
\alias{isStringANumber}
\title{Identify if a string is a number}
\usage{
isStringANumber(x, decimalSeparator = ".")
}
\arguments{
\item{x}{\code{character} vector.}

\item{decimalSeparator}{\code{character} used as a decimal separator ("," or "."); \code{decimalSeparator = "."} by default.}
}
\value{
Return a \code{logical} vector indicating, for each element of \code{x}, if the string correspond to a number.
}
\description{
Check if elements of a \code{character} vector are numbers, in their scientific notation or not.
}
\examples{
x <- c("1234", "12.34","1.1E+10", "12,34", "1,1E+10", "123A")
isStringANumber(x,".")
isStringANumber(x,",")
}
