% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isStringCanBeADate.R
\name{isStringCanBeADate}
\alias{isStringCanBeADate}
\title{Identify if a string can be converted into a date}
\usage{
isStringCanBeADate(x, origin = as.Date("1899-12-30"),
  format = ("\%Y-\%m-\%d"), warning = FALSE)
}
\arguments{
\item{x}{vecteur, \code{matrix} ou \code{data.frame} de \code{character} ou \code{numeric}.}
}
\value{
Return a \code{logical} vector indicating, for each element of \code{x}, if the string can be converted into a \code{Date}.
}
\description{
Check if a vector, \code{matrix} or \code{data.frame} of class \code{numeric} or \code{character} can be converted into a \code{Date}.
}
\details{
String representing valid dates, according to the specified format, numbers and \code{NA} (but not "NA") are considered to be convertible into a \code{Date}.
}
\examples{
#Examples of valid cases
x <- c("2019-06-20", 43636, "43636.1", "43636,1", NA)
isStringCanBeADate(x)

#Examples of non-valid cases
y<- c("AA", "2012/05/02", "06-05-2012", "2019-06-32", "NA")
isStringCanBeADate(y)

}
