% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isStringSciNumber.R
\name{isStringSciNumber}
\alias{isStringSciNumber}
\title{Identify if a string is a scientific number}
\usage{
isStringSciNumber(x, decimalSeparator = ".")
}
\arguments{
\item{x}{\code{character} vector.}

\item{decimalSeparator}{\code{character} used as a decimal separator ("," or "."); \code{decimalSeparator = "."} by default.}
}
\value{
Return a \code{logical} vector indicating, for each element of \code{x}, if the string correspond to a number in scientific notation.
}
\description{
Check if elements of a \code{character} vector are numbers expressed in scientific notation.
}
\details{
The function can identify positive or negative numbers, integer or decimals, followed by \code{E} (or \code{e}), another symbol (\code{+}, \code{-} or nothing) and digits representing the exponent.
}
\examples{
x <- c("1234", "-1","+1", "12.34","12,34", "123A","-1.1","E+1")
y <- c("1E+1","-1e-1", "+1.1e-1","1.1E10","1,1e+10")
isStringSciNumber(x)
isStringSciNumber(y)
isStringSciNumber(y, decimalSeparator = ",")

}
